refactor: repo-wide clean-code pass and public API evolution
A smell audit of every main source (Refactoring Guru vocabulary), followed by a
behavior-preserving fix pass and an authorized public-API pass, all verified by
the full build: __.compile with -Werror -Wunused:all, every module and example
test suite, scalafmt and scalafix gates.

Public API (old forms kept as deprecated delegates where cheap):
- widgets: ChartOptions/DataTableOptions/ScrollbarOptions/OrbitSpinnerOptions
  parameter objects replace 12-16-parameter constructor lists; Bounds type
  replaces (Double, Double) world-bounds tuples; typed-row DataTable[K] with
  KeyedRow/selectedKey/selectKey so selection tracks identity across the internal
  sort (fromStrings keeps text rows working); DropdownState owns the navigation
  arithmetic the dsl used to reach into.
- core: Buffer.diff takes the TrailingCellPolicy ADT instead of a Boolean;
  Modifiers/KeyModifiers share the private FlagBits kernel; Layout.split2..5
  consolidate over a private splitter, tuple signatures unchanged.
- runtime: Signal change detection goes through the SignalEquality typeclass
  (default == semantics, Double/Float total order).
- terminal: JLine3Backend split into cohesive collaborators (TerminalDressing,
  EventPump, ReplyQueries, ScrollRegions, TitleStack); locking unchanged.
- test-support: Pilot mouse methods gain Position overloads.
- macros: the Double field given is renamed to double, matching Field.double.

Private refactors in every module: Buffer's diff engine extracted to FrameDiff,
the memoized style transform shared as MemoizedStyleTransform, KeyEvent.keyCodeFor
as a lookup table, Paragraph's WrapWalker class, TerminalRunner's loop body and a
DispatchOutcome enum replacing a Boolean branch API, FieldBinding.parsed, shared
ProgressMeterBuilders[Self] in the dsl, and magic numbers named throughout.

Examples: procmon rewritten to keyed rows (the parse-the-pid-back-out hack is
gone), airsensor's Metric gets honest identity semantics and Band.cascade takes
validated ordered thresholds, an interrupted sensor read now restores the
thread's interrupt status (with a regression test), weather's JSON parser loops
consolidate into one parameterized loop, and locale-sensitive formatting moves
to Locale.ROOT/fixed().

+++ macros/src/main/scala/io/worxbend/tui/macros/Field.scala
@@ -12,12 +12,9 @@ final case class Field[A](spec: FieldSpec, parse: String => Either[String, A]):
 
   /** Runs `f` on every successfully parsed value, leaving parse failures untouched.
     *
-    * Use `map` to normalise a value — trim it, lower-case it — not to change what it is. Building a validator from the
-    * wrong factory for the field's declared type (a `Field.text` for an `Int` field) is now caught where the form is
-    * declared, because [[FieldSpec.input]] records which factory made it. One case survives that check: `map` keeps the
-    * spec it was called on, so `Field.int("age").map(_.toString)` still claims to be an `IntField` while producing a
-    * `String`. The assembled values reach the case class's constructor with no type check, so that one fails on submit
-    * with a `ClassCastException` pointing at the constructor rather than at this call.
+    * Use `map` to normalise a value — trim it, lower-case it — not to change what it is. `map` keeps the spec it was
+    * called on, so changing the type here falls under the residual `ClassCastException` caveat documented on the
+    * [[Field$ companion object]].
     */
   def map[B](f: A => B): Field[B] =
     mapValidated(value => Right(f(value)))
@@ -25,8 +22,9 @@ final case class Field[A](spec: FieldSpec, parse: String => Either[String, A]):
   /** Chains a validation step onto the parser: `f` may reject a parsed value by returning `Left(message)`, and that
     * message is what the form shows next to the field.
     *
-    * The same residual caveat as [[map]] applies — when this field validates a derived [[FormSpec]], `B` must remain
-    * the case class's declared field type. `mapValidated` is for rejecting values, not for changing their type.
+    * `mapValidated` is for rejecting values, not for changing their type — when this field validates a derived
+    * [[FormSpec]], `B` must remain the case class's declared field type, or the residual `ClassCastException` caveat
+    * documented on the [[Field$ companion object]] applies.
     */
   def mapValidated[B](f: A => Either[String, B]): Field[B] =
     Field(spec, raw => parse(raw).flatMap(f))
@@ -43,8 +41,13 @@ final case class Field[A](spec: FieldSpec, parse: String => Either[String, A]):
   * control: `int` and `long` are both the whole-number field, `double` and `bigDecimal` are both the decimal one, and
   * `text`, `uuid`, `localDate`, `localTime`, `localDateTime` and `duration` are all the plain text field. So attaching
   * a `Field.int("count")` validator to a field the case class declares as `Long` passes that check and instead fails on
-  * submit with a `ClassCastException` from the case class's constructor — the same residual case [[Field.map]]
-  * describes. Pick the factory named after the field's declared type, not after the control it happens to render as.
+  * submit with a `ClassCastException` from the case class's constructor. Pick the factory named after the field's
+  * declared type, not after the control it happens to render as.
+  *
+  * The same residual case can also be reached through [[Field.map]] and [[Field.mapValidated]]: both keep the spec they
+  * were called on, so `Field.int("age").map(_.toString)` still claims to be an `IntField` while producing a `String`.
+  * The assembled values reach the case class's constructor with no type check, so that one fails on submit with a
+  * `ClassCastException` pointing at the constructor rather than at the offending call.
   */
 object Field:
 
@@ -52,7 +55,7 @@ object Field:
 
   def int(name: String): Field[Int] = FormFieldType.int.field(name)
 
-  def double(name: String): Field[Double] = FormFieldType.decimal.field(name)
+  def double(name: String): Field[Double] = FormFieldType.double.field(name)
 
   def bool(name: String): Field[Boolean] = FormFieldType.bool.field(name)
